==> 03_nsduh_estimates.Rmd <==
---
title: "Extract model estimates"
subtitle: "Depression among US adults at the intersection of race/ethnicity, sex/gender, and sexual orientation: A design-weighted intersectional MAIHDA"
author: "F. Hunter McGuire, PhD, MPH"
date: "December 23, 2024"
output:
  html_document:
    toc: yes
    toc_float: yes
    toc-location: left
    code_folding: show
  pdf_document:
    toc: yes
editor_options: 
  chunk_output_type: console
---

# Purpose

The purpose of this file is to summarize estimates from design-weighted Bayesian multilevel models fit with the package `csSampling` (backend `brms` and `rstan`). These models were fit in a previous file named `02_nsduh_analysis.R`.

## Model specifications

Individuals (level 1) were nested within intersectional groups (level 2) defined by race/ethnicity, sex/gender, and sexual orientation. Models were fit with Markov chain Monte Carlo (MCMC) with a Bernoulli outcome distribution and logit link. We specified weakly informative priors with variance parameters having an inverse gamma distribution (alpha=2, beta=2) and variable coefficients on the logit scale having a Gaussian distribution (mean=0, standard deviation=2). Four Markov chains were run for 4,000 iterations including 2,000 warmups (total post-warmup iterations per model = 8,000). 

For each outcome (i.e., lifetime MDE and past-year MDE), the baseline model was an age-adjusted intersectional model (i.e., age category as the only level 1 covariate) with random intercepts for intersectional groups. We then estimated an age-adjusted intersectional interaction model (Full model) with age category and all three identity variables (race/ethnicity, sex/gender, sexual orientation) as level 1 covariates. 

# Setup

The first code chunk loads the necessary `R` packages, Stan model fits, and NSDUH 2015-2020 (combined) data frame.

I then extract the parameter estimates from each model, including the effective sample size (ESS) and Gelman-Rubin statistic (r-hat) that will later be used to evaluate Markov chain Monte Carlo (MCMC) model convergence.

Finally, I load the NSDUH 2015-2020 dataset.

```{r setup, message=FALSE, warning=FALSE}

# set global options for how chunks are processed/output
knitr::opts_chunk$set(message=FALSE, warning=FALSE)

# Load packages
library(tidyverse) # Load `tidyverse` suite of packages
library(plyr) # Tools for Splitting, Applying and Combining Data
library(rstan) # R interface for Stan
library(bayesplot) # Plotting for Bayesian models
library(rstudioapi) # Retrieve information about an RStudio Editor
library(survey) # design-weighted analysis
library(ggthemes)

# set working directory to source file location
### get full file path
RMDpath <- rstudioapi::getSourceEditorContext()$path
### remove file name from file path
RMDpath <- gsub('/03_nsduh_estimates.Rmd','', RMDpath)
### set the working directory
setwd(RMDpath)

# Load NSDUH data frame (2015-2020 combined data collection years)
nsduh <- readRDS("nsduh.RDS")

# Load helper functions used to extract/manipulate model estimates
source("00_helper_functions.R")
```

# Define intersectional groups

Here I define intersectional groups defined by race/ethnicity, sex/gender, and sexual orientation. This produces 42 mutually-exclusive groups based on 7 races/ethnicity, 2 sex/gender, and 3 sexual orientation categories.

```{r}
# Get strata counts (N) and merge with groupName
# Past-year MDE
group_count_pyMDE <- nsduh %>% 
  filter(sp_2 == 1) %>% # subpopulation indicator
  count('strata') %>% 
  mutate(groupID = strata) %>% 
  cbind(groupName, groupNameShort) %>% 
  select(groupID, groupName, groupNameShort, freq)
# Lifetime MDE
group_count_ltMDE <- nsduh %>% 
  filter(sp_1 == 1) %>% # subpopulation indicator
  count('strata') %>% 
  mutate(groupID = strata) %>% 
  cbind(groupName, groupNameShort) %>% 
  select(groupID, groupName, groupNameShort, freq)

# Create combined counts table
group_count_overall <- group_count_pyMDE %>% 
  mutate(n_pyMDE = freq) %>% 
  mutate(n_ltMDE = group_count_ltMDE$freq) %>% 
  select(groupID, groupName, n_ltMDE, n_pyMDE)

# Create combined counts table
group_count_overall <- group_count_pyMDE %>% 
  mutate(n_pyMDE = freq) %>% 
  mutate(n_ltMDE = group_count_ltMDE$freq) %>% 
  select(groupID, groupName, n_ltMDE, n_pyMDE)

group_count_year <- nsduh %>% 
  filter(sp_1 == 1) %>% # subpopulation indicator
  count(c("strata", "year")) %>% 
  pivot_wider(names_from = "year", values_from = "freq") %>% 
  rowwise() %>%
  mutate(overall = sum(c(`2015`, `2016`, `2017`, `2018`, `2019`, `2020`), 
                       na.rm = TRUE),
         groupID = strata) %>% 
  replace_na(list(`2020` = 0)) %>% 
  cbind(groupName, groupNameShort) %>% 
  select(groupName, groupID, overall, `2015`, `2016`, 
         `2017`, `2018`, `2019`, `2020`)
```

# AIM 1: Past-year MDE

This code section works with models estimating past-year major depressive episodes (MDE).

## Load model fits: Past-year MDE

```{r}
## Load model fits (produced via "nsduh_pyMDE_analysis.R")
##### Note: Ensure model fits are saved in a '02_fits/' folder within your
#####       current working directory.
pyMDE_m1 <- readRDS("02_fits/weighted_csSampling/pyMDE_m1.RDS")
pyMDE_m2 <- readRDS("02_fits/weighted_csSampling/pyMDE_m2.RDS")

## Extract stan_fit objects
## CAUTION: The `stan_fit` objects are NOT variance-adjusted to correct for
##          dependencies between observations in the NSDUH sampling design.
##          The `stan_fit` objects are survey-weighted posterior distributions 
##          that correct results for unequal probabilities of selection and
##          differential response rates to the survey. 
pyMDE_m1 <- pyMDE_m1$stan_fit
pyMDE_m2 <- pyMDE_m2$stan_fit

## Extract parameter estimates, effective sample size (ESS), and r-hat 
## from each model fit. Specify 95% credible intervals.
pyMDE_m1_sum <- as.data.frame(rstan::summary(pyMDE_m1,
                                      probs = c(0.025, 0.975))$summary)
pyMDE_m2_sum <- as.data.frame(rstan::summary(pyMDE_m2,
                                      probs = c(0.025, 0.975))$summary)
```

## Variance estimates for all models

```{r}
# Summarize SD, variance, VPC, and PCV
pyMDE_var <- variance(models = c(pyMDE_m1, pyMDE_m2))
```

## Baseline model: Age-adjusted intersectional model (null)

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(pyMDE_m1)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Baseline model: Age-adjusted intersectional") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# Rename parameter labels and save the model fit object as an array
# This is done to aid chart comprehension
pyMDE_m1_array <- renameArray(model = pyMDE_m1,
                              names = groupName,
                              start = 48,
                              finish = 89)
# 10/24/2024 Note: this code was updated to capture the correct intercept,
# which is "b_Intercept".
# In a prior version of this code, this was incorrectly identified as the 
# parameter "Intercept" -- the incorrect "Intercept" parameter is mean-centered
# and is much lower than the correct "b_Intercept" parameter, which is on the
# original scale.
# Therefore, this produced lower prevalence estimates for each intersectional
# group in our original analysis and publication.
dimnames(pyMDE_m1_array)[[3]][c(1:3, 5, 91)] <-
  c("age_2534", "age_3549", "age_50plus", 
    "Between-group std. dev.", "b_Intercept")

# define parameters to display in plots
parameters <- c(c(dimnames(pyMDE_m1_array)[[3]][c(1:3, 5, 91)], 
                  groupName[1:42]))

# Autocorrelation plots
autocor_plots(array = pyMDE_m1_array, 
              parsList = parameters)

# Trace plots for each parameter
trace_plots(array = pyMDE_m1_array,
            parsList = parameters)
```

### Model estimates

Age-adjusted prevalence

```{r}
# Define population age distribution weights
age_weights <- c(0.138, 0.159, 0.246, 0.456)
# Note: these percentages come from Table 1
# 18-25 = 75,591 (13.8%)
# 26-34 = 48,242 (15.9%)
# 35-49 = 62,110 (24.6%)
# 50+   = 48,779 (45.6%)

# Generate age-standardized prevalence
pyMDE_prevalence <-
  age_adj_prevalence(model = pyMDE_m1, 
                     # 10/24/2024 Note: We have updated this to the correct
                     # intercept parameter ("b_Intercept")
                     parameters =  c("b_Intercept", "b[1]", "b[2]", "b[3]"),
                     age_weights = age_weights)
```

Coefficients

```{r}
# Get ORs and CIs
pyMDE_m1_estimates <- pyMDE_m1_sum %>% 
  # keep rows with beta coefficients
  # 10/24/2024 Note: We have updated this to the correct intercept parameter 
  # ("b_Intercept")
  filter(rownames(.) %in% c("b[1]", "b[2]", "b[3]", "b_Intercept")) %>% 
  mutate(
    # Create a new column named "parameter" to store parameter names
    parameter = c("26-34",
                  "35-49",
                  "50+",
                  "Intercept"),
    # Calculate odds ratios and 95% credible intervals
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

## Full model: Fully-adjusted intersectional interaction

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(pyMDE_m2)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Full model: Intersectional interaction") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# Rename parameter labels
# 10/24/2024 Note: this code was updated to capture the correct intercept,
# which is "b_Intercept".
# In a prior version of this code, this was incorrectly identified as the 
# parameter "Intercept" -- the incorrect "Intercept" parameter is mean-centered
# and is much lower than the correct "b_Intercept" parameter, which is on the
# original scale.
# Therefore, this produced lower prevalence estimates for each intersectional
# group in our original analysis and publication.
pyMDE_m2_array <- as.array(pyMDE_m2)
dimnames(pyMDE_m2_array)[[3]][c(1:12, 14, 100)] <-
  c("age_2534", "age_3549", "age_50plus",
    "Asian", "Black", "Hispanic/Latine", 
    "NAAN", "NHPI", "Multiracial",
    "Woman", "Gay/Lesbian", "Bisexual",
    "Between-group std. dev.", "b_Intercept")


# define parameters to display in plots
parameters <- dimnames(pyMDE_m2_array)[[3]][c(1:12, 14, 100)]

# Autocorrelation plots
autocor_plots(array = pyMDE_m2_array, 
              parsList = parameters)

# Trace plots for each parameter
trace_plots(array = pyMDE_m2_array,
            parsList = parameters)
```

### Model estimates

```{r}
# Get ORs and CIs
pyMDE_m2_estimates <- pyMDE_m2_sum %>% 
  # keep rows with beta coefficients
  # 10/24/2024 Note: We have updated this to the correct intercept parameter 
  # ("b_Intercept")
  filter(rownames(.) %in% c("b[1]", "b[2]", "b[3]", 
                            "b[4]", "b[5]", "b[6]",
                            "b[7]", "b[8]", "b[9]",
                            "b[10]", "b[11]", "b[12]",
                            "b_Intercept")) %>% 
  mutate(
    parameter = c("26-34",
                  "35-49",
                  "50+",
                  "Asian",
                  "Black/African-American",
                  "Hispanic/Latine",
                  "Native American / Alaskan Native",
                  "Native Hawaiian / Pacific Islander",
                  "Multiracial",
                  "Woman",
                  "Gay/Lesbian",
                  "Bisexual",
                  "Intercept"),
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

### Total, main, and intersectional effects

```{r}
pyMDE_effects <- 
  calculate_effects(model = pyMDE_m2,
                    # 10/24/2024 Note: Updated to have the correct intercept
                    # parameter ("b_Intercept").
                    parameters = c("b_Intercept", "b[1]", "b[2]", "b[3]",
                                    "b[4]", "b[5]", "b[6]", "b[7]", "b[8]", 
                                    "b[9]", "b[10]", "b[11]", "b[12]"),
                    age_weights = age_weights)

# Step 4: Combine all estimated effects into a single table
pyMDE_effect_summary <-
  # start with total effects
  pyMDE_effects$total_sum %>% 
  select(`PY Total Effects (95% CI)` = Combined) %>% 
  # add main effects
  cbind(., 
        pyMDE_effects$main_sum %>% 
          select(`PY Main Effects (95% CI)` = Combined)) %>% 
  cbind(., 
        pyMDE_effects$intersectional_sum %>% 
          select(`PY % Point Diff.` = Combined)) %>% 
  cbind(.,
        pyMDE_effects$intersectional_pct_sum %>% 
          select(`PY % Diff.` = Combined)) %>% 
  # add intersectional group names and group IDs
  cbind(`Intersectional Group` = groupName, `Group ID` = c(1:42), .)
```


# AIM 2: Lifetime MDE

This code section works with models estimating lifetime major depressive episodes (MDE).

## Load model fits: Lifetime MDE

```{r}
## Load model fits (produced via "nsduh_ltMDE_analysis.R")
##### Note: Ensure model fits are saved in a 'fits/' folder within your
#####       current working directory.
ltMDE_m1 <- readRDS("02_fits/weighted_csSampling/ltMDE_m1.RDS")
ltMDE_m2 <- readRDS("02_fits/weighted_csSampling/ltMDE_m2.RDS")

## Extract stan_fit objects
## CAUTION: The `stan_fit` objects are NOT variance-adjusted to correct for
##          dependencies between observations in the NSDUH sampling design.
##          The `stan_fit` objects are survey-weighted posterior distributions 
##          that correct results for unequal probabilities of selection and
##          differential response rates to the survey. 
ltMDE_m1 <- ltMDE_m1$stan_fit
ltMDE_m2 <- ltMDE_m2$stan_fit

## Extract parameter estimates, effective sample size (ESS), and r-hat 
## from each model fit. Specify 95% credible intervals.
ltMDE_m1_sum <- as.data.frame(rstan::summary(ltMDE_m1,
                                     probs = c(0.025, 0.975))$summary)
ltMDE_m2_sum <- as.data.frame(rstan::summary(ltMDE_m2,
                                     probs = c(0.025, 0.975))$summary)
```

## Variance estimates for all models

```{r}
# Summarize SD, variance, VPC, and PCV
ltMDE_var <- variance(models = c(ltMDE_m1, ltMDE_m2))
```

## Baseline model: Age-adjusted intersectional model (null)

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(ltMDE_m1)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Baseline model: Age-adjusted intersectional") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# Rename parameter labels save the model fit object as an array
# This is done to aid chart comprehension
ltMDE_m1_array <- renameArray(model = ltMDE_m1,
                             names = groupName,
                             start = 48,
                             finish = 89)
# 10/24/2024 Note: this code was updated to capture the correct intercept,
# which is "b_Intercept".
# In a prior version of this code, this was incorrectly identified as the 
# parameter "Intercept" -- the incorrect "Intercept" parameter is mean-centered
# and is much lower than the correct "b_Intercept" parameter, which is on the
# original scale.
# Therefore, this produced lower prevalence estimates for each intersectional
# group in our original analysis and publication.
dimnames(ltMDE_m1_array)[[3]][c(1:3, 5, 91)] <-
  c("age_2534", "age_3549", "age_50plus", 
    "Between-group std. dev.", "b_Intercept")

# define parameters to display in plots
parameters <- c(dimnames(ltMDE_m1_array)[[3]][c(1:3, 5, 91)],
                groupName[1:42])

# Autocorrelation plots
autocor_plots(array = ltMDE_m1_array, 
              parsList = parameters)

# Trace plots for each parameter
trace_plots(array = ltMDE_m1_array,
            parsList = parameters)
```

### Model estimates

Age-standardized prevalence

```{r}
# Define population age distribution weights
age_weights <- c(0.138, 0.159, 0.246, 0.456)
# Note: these percentages come from Table 1
# 18-25 = 75,591 (13.8%)
# 26-34 = 48,242 (15.9%)
# 35-49 = 62,110 (24.6%)
# 50+   = 48,779 (45.6%)

# Generate age-standardized prevalence
ltMDE_prevalence <-
  age_adj_prevalence(model = ltMDE_m1, 
                     # 10/24/2024 Note: We have updated this to the correct
                     # intercept parameter ("b_Intercept")
                     parameters =  c("b_Intercept", "b[1]", "b[2]", "b[3]"),
                     age_weights = age_weights)
```

Model coefficients

```{r}
# Get ORs and CIs
ltMDE_m1_estimates <- ltMDE_m1_sum %>% 
  # keep rows with beta coefficients
  # 10/24/2024 Note: We have updated this to the correct intercept parameter 
  # ("b_Intercept")
  filter(rownames(.) %in% c("b[1]", "b[2]", "b[3]", "b_Intercept")) %>% 
  mutate(
    # Create a new column named "parameter" to store parameter names
    parameter = c("26-34",
                  "35-49",
                  "50+",
                  "Intecept"),
    # Calculate odds ratios with 95% CIs
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

## Full model: Fully-adjusted intersectional interaction

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(ltMDE_m2)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Full model: Intersectional interaction") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# save the model fit object as an array
ltMDE_m2_array <- as.array(ltMDE_m2)
# Rename parameter labels
# 10/24/2024 Note: this code was updated to capture the correct intercept,
# which is "b_Intercept".
# In a prior version of this code, this was incorrectly identified as the 
# parameter "Intercept" -- the incorrect "Intercept" parameter is mean-centered
# and is much lower than the correct "b_Intercept" parameter, which is on the
# original scale.
# Therefore, this produced lower prevalence estimates for each intersectional
# group in our original analysis and publication.
dimnames(ltMDE_m2_array)[[3]][c(1:12, 14, 100)] <-
  c("age_2534", "age_3549", "age_50plus",
    "Asian", "Black", "Hispanic/Latine", 
    "NAAN", "NHPI", "Multiracial",
    "Woman", "Gay/Lesbian", "Bisexual",
    "Between-group std. dev.", "b_Intercept")
# define parameters to display in plots
parameters <- c(dimnames(ltMDE_m2_array)[[3]][c(1:12, 14, 100)])

# Autocorrelation plots
autocor_plots(array = ltMDE_m2_array, 
              parsList = parameters)

# Trace plots
trace_plots(array = ltMDE_m2_array,
            parsList = parameters)
```

### Model estimates

```{r}
# Get ORs and CIs
ltMDE_m2_estimates = ltMDE_m2_sum %>% 
  # keep rows with beta coefficients
  # 10/24/2024 Note: We have updated this to the correct intercept parameter 
  # ("b_Intercept")
  filter(rownames(.) %in% c("b[1]", "b[2]", "b[3]", 
                            "b[4]", "b[5]", "b[6]",
                            "b[7]", "b[8]", "b[9]",
                            "b[10]", "b[11]", "b[12]",
                            "b_Intercept")) %>% 
  mutate(
    parameter = c("26-34",
                  "35-49",
                  "50+",
                  "Asian",
                  "Black/African-American",
                  "Hispanic/Latine",
                  "Native American / Alaskan Native",
                  "Native Hawaiian / Pacific Islander",
                  "Multiracial",
                  "Woman",
                  "Gay/Lesbian",
                  "Bisexual",
                  "Intercept"),
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

### Total, main, and intersectional effects

```{r}
ltMDE_effects <- 
  calculate_effects(model = ltMDE_m2,
                    # 10/24/2024 Note: Updated to have the correct intercept
                    # parameter ("b_Intercept").
                    parameters = c("b_Intercept", "b[1]", "b[2]", "b[3]",
                                    "b[4]", "b[5]", "b[6]", "b[7]", "b[8]", 
                                    "b[9]", "b[10]", "b[11]", "b[12]"),
                    age_weights = age_weights)

# Combine all estimated effects into a single table
ltMDE_effect_summary <-
  # start with total effects
  ltMDE_effects$total_sum %>% 
  select(`LT Total Effects (95% CI)` = Combined) %>% 
  # then add main effects
  cbind(., 
        ltMDE_effects$main_sum %>% 
          select(`LT Main Effects (95% CI)` = Combined)) %>% 
  # then add intersectional effects
  cbind(., 
        ltMDE_effects$intersectional_sum %>% 
          select(`LT % Point Diff.` = Combined)) %>% 
  cbind(.,
        ltMDE_effects$intersectional_pct_sum %>% 
          select(`LT % Diff.` = Combined))  %>%  
  # add intersectional group names and group IDs
  cbind(`Intersectional Group` = groupName, `Group ID` = c(1:42), .)
```


# Save output

This output will be used to create table and figures in the file `05_tables_figures.Rmd`.

```{r}
# intersectional group counts
saveRDS(group_count_overall,
        "03_tables_figures/model_estimates/group_count_overall.RDS")
saveRDS(group_count_year,
        "03_tables_figures/model_estimates/group_count_year.RDS")

# prevalence estimates
saveRDS(pyMDE_prevalence,
        "03_tables_figures/model_estimates/pyMDE_prevalence.RDS")
saveRDS(ltMDE_prevalence,
        "03_tables_figures/model_estimates/ltMDE_prevalence.RDS")

# model fit summaries
saveRDS(pyMDE_m1_sum,
        "03_tables_figures/model_estimates/pyMDE_m1_sum.RDS")
saveRDS(pyMDE_m2_sum,
        "03_tables_figures/model_estimates/pyMDE_m2_sum.RDS")
saveRDS(ltMDE_m1_sum,
        "03_tables_figures/model_estimates/ltMDE_m1_sum.RDS")
saveRDS(ltMDE_m2_sum,
        "03_tables_figures/model_estimates/ltMDE_m2_sum.RDS")
saveRDS(pyMDE_m1_estimates,
        "03_tables_figures/model_estimates/pyMDE_m1_estimates.RDS")
saveRDS(pyMDE_m2_estimates,
        "03_tables_figures/model_estimates/pyMDE_m2_estimates.RDS")
saveRDS(ltMDE_m1_estimates,
        "03_tables_figures/model_estimates/ltMDE_m1_estimates.RDS")
saveRDS(ltMDE_m2_estimates,
        "03_tables_figures/model_estimates/ltMDE_m2_estimates.RDS")
# variance
saveRDS(pyMDE_var,
        "03_tables_figures/model_estimates/pyMDE_var.RDS")
saveRDS(ltMDE_var,
        "03_tables_figures/model_estimates/ltMDE_var.RDS")

# intersectional effects (excess/reduced prevalence)
saveRDS(pyMDE_effect_summary,
        "03_tables_figures/model_estimates/pyMDE_effect_summary.RDS")
saveRDS(ltMDE_effect_summary,
        "03_tables_figures/model_estimates/ltMDE_effect_summary.RDS")
```

# Sensitivity analysis

## Remove prior files

```{r}
rm(list=ls()) 
```

## Load functions and data

```{r}
# Load NSDUH data frame (2015-2020 combined data collection years)
nsduh <- readRDS("nsduh.RDS")

# Load helper functions used to extract/manipulate model estimates
source("00_helper_functions.R")
```


# AIM 1: Past-year MDE

This code section works with models estimating past-year major depressive episodes (MDE).

## Load model fits: Past-year MDE

```{r}
# Load model fits
pyMDE_m1 <- 
  readRDS(file = "02_fits/unweighted_brms/pyMDE_m1_unweighted.rds")
pyMDE_m2 <- 
  readRDS(file = "02_fits/unweighted_brms/pyMDE_m2_unweighted.rds")

# Extract rstan fits
pyMDE_m1 <- pyMDE_m1$fit
pyMDE_m2 <- pyMDE_m2$fit

## Extract parameter estimates, effective sample size (ESS), and r-hat 
## from each model fit. Specify 95% credible intervals.
pyMDE_m1_sum = as.data.frame(rstan::summary(pyMDE_m1,
                                probs = c(0.025, 0.975))$summary)
pyMDE_m2_sum = as.data.frame(rstan::summary(pyMDE_m2,
                                probs = c(0.025, 0.975))$summary)
```

## Variance estimates for all models

```{r}
# Summarize SD, variance, VPC, and PCV
pyMDE_var <- variance(models = c(pyMDE_m1, pyMDE_m2))
```

## Baseline model: Age-adjusted intersectional model (null)

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(pyMDE_m1)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Baseline model: Age-adjusted intersectional") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# Rename parameter labels 
# This is done to aid chart comprehension
pyMDE_m1_array <- renameArray(model = pyMDE_m1,
                              names = groupName,
                              start = 6,
                              finish = 47)
dimnames(pyMDE_m1_array)[[3]][1] = "Intercept"
dimnames(pyMDE_m1_array)[[3]][2] = "age_2534"
dimnames(pyMDE_m1_array)[[3]][3] = "age_3549"
dimnames(pyMDE_m1_array)[[3]][4] = "age_50plus"
dimnames(pyMDE_m1_array)[[3]][5] = "Between-group std. dev."

# define parameters to display in plots
parameters <- c(c(dimnames(pyMDE_m1_array)[[3]][1:5], 
                  groupName[1:42]))

# Autocorrelation plots
autocor_plots(array = pyMDE_m1_array, 
              parsList = parameters)

# Trace plots for each parameter
trace_plots(array = pyMDE_m1_array,
            parsList = parameters)
```

### Model estimates

Age-adjusted prevalence

```{r}
# Define population age distribution weights
age_weights <- c(0.138, 0.159, 0.246, 0.456)
# Note: these percentages come from Table 1
# 18-25 = 75,591 (13.8%)
# 26-34 = 48,242 (15.9%)
# 35-49 = 62,110 (24.6%)
# 50+   = 48,779 (45.6%)

# Generate age-standardized prevalence
pyMDE_prevalence <-
  age_adj_prevalence(model = pyMDE_m1, 
                     parameters =  c("b_Intercept", "b_age2634", 
                                     "b_age3549", "b_age50plus"),
                     age_weights = age_weights)
```

Coefficients

```{r}
# Get ORs and CIs
pyMDE_m1_estimates <- pyMDE_m1_sum %>% 
  # keep rows with beta coefficients
  filter(row_number() %in% c(1:4)) %>% 
  mutate(
    parameter = c("Intercept",
                  "26-34",
                  "35-49",
                  "50+"),
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

## Full model: Fully-adjusted intersectional interaction

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(pyMDE_m2)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Full model: Intersectional interaction") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# Rename parameter labels
pyMDE_m2_array <- as.array(pyMDE_m2)
dimnames(pyMDE_m2_array)[[3]][1] = "Intercept"
dimnames(pyMDE_m2_array)[[3]][2] = "age_2534"
dimnames(pyMDE_m2_array)[[3]][3] = "age_3549"
dimnames(pyMDE_m2_array)[[3]][4] = "age_50plus"
dimnames(pyMDE_m2_array)[[3]][5] = "Asian"
dimnames(pyMDE_m2_array)[[3]][6] = "Black"
dimnames(pyMDE_m2_array)[[3]][7] = "Hispanic/Latine"
dimnames(pyMDE_m2_array)[[3]][8] = "NAAN"
dimnames(pyMDE_m2_array)[[3]][9] = "NHPI"
dimnames(pyMDE_m2_array)[[3]][10] = "Multiracial"
dimnames(pyMDE_m2_array)[[3]][11] = "Woman"
dimnames(pyMDE_m2_array)[[3]][12] = "Gay/Lesbian"
dimnames(pyMDE_m2_array)[[3]][13] = "Bisexual"
dimnames(pyMDE_m2_array)[[3]][14] = "Between-group std. dev."

# define parameters to display in plots
parameters <- dimnames(pyMDE_m2_array)[[3]][1:14]

# Autocorrelation plots
autocor_plots(array = pyMDE_m2_array, 
              parsList = parameters)

# Trace plots for each parameter
trace_plots(array = pyMDE_m2_array,
            parsList = parameters)
```

### Model estimates

```{r}
# Get ORs and CIs
pyMDE_m2_estimates <- pyMDE_m2_sum %>% 
  # keep rows with beta coefficients
  filter(row_number() %in% c(1:13)) %>% 
  mutate(
    parameter = c("Intercept",
                  "26-34",
                  "35-49",
                  "50+",
                  "Asian",
                  "Black/African-American",
                  "Hispanic/Latine",
                  "Native American / Alaskan Native",
                  "Native Hawaiian / Pacific Islander",
                  "Multiracial",
                  "Woman",
                  "Gay/Lesbian",
                  "Bisexual"),
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

### Total, main, and intersectional effects

```{r}
pyMDE_effects <- 
  calculate_effects(model = pyMDE_m2,
                    parameters = c("b_Intercept", "b_age2634", 
                                    "b_age3549", "b_age50plus",
                                    "b_asian", "b_black",
                                    "b_hispanic", "b_naan",
                                    "b_nhpi", "b_multi",
                                    "b_woman", "b_gay",
                                    "b_bisexual"),
                    age_weights = age_weights)

# Combine all estimated effects into a single table
pyMDE_effect_summary <-
  # start with total effects
  pyMDE_effects$total_sum %>% 
  select(`PY Total Effects (95% CI)` = Combined) %>% 
  # add main effects
  cbind(., 
        pyMDE_effects$main_sum %>% 
          select(`PY Main Effects (95% CI)` = Combined)) %>% 
  cbind(., 
        pyMDE_effects$intersectional_sum %>% 
          select(`PY % Point Diff.` = Combined)) %>% 
  cbind(.,
        pyMDE_effects$intersectional_pct_sum %>% 
          select(`PY % Diff.` = Combined)) %>% 
  # add intersectional group names and group IDs
  cbind(`Intersectional Group` = groupName, `Group ID` = c(1:42), .)
```


# AIM 2: Lifetime MDE

This code section works with models estimating lifetime major depressive episodes (MDE).

## Load model fits: Lifetime MDE

```{r}
## Load model fits (produced via "nsduh_ltMDE_analysis.R")
ltMDE_m1 <- 
  readRDS(file = "02_fits/unweighted_brms/ltMDE_m1_unweighted.rds")
ltMDE_m2 <- 
  readRDS(file = "02_fits/unweighted_brms/ltMDE_m2_unweighted.rds")

## Extract stan_fit objects
ltMDE_m1 = ltMDE_m1$fit
ltMDE_m2 = ltMDE_m2$fit

## Extract parameter estimates, effective sample size (ESS), and r-hat 
## from each model fit. Specify 95% credible intervals.
ltMDE_m1_sum = as.data.frame(summary(ltMDE_m1,
                                probs = c(0.025, 0.975))$summary)
ltMDE_m2_sum = as.data.frame(summary(ltMDE_m2,
                                probs = c(0.025, 0.975))$summary)
```

## Variance estimates for all models

```{r}
# Summarize SD, variance, VPC, and PCV
ltMDE_var <- variance(models = c(ltMDE_m1, ltMDE_m2))
```

## Baseline model: Age-adjusted intersectional model (null)

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(ltMDE_m1)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Baseline model: Age-adjusted intersectional") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# Rename parameter labels 
# This is done to aid chart comprehension
ltMDE_m1_array = renameArray(model = ltMDE_m1,
                             names = groupName,
                             start = 6,
                             finish = 47)
dimnames(ltMDE_m1_array)[[3]][1] = "Intercept"
dimnames(ltMDE_m1_array)[[3]][2] = "age_2534"
dimnames(ltMDE_m1_array)[[3]][3] = "age_3549"
dimnames(ltMDE_m1_array)[[3]][4] = "age_50plus"
dimnames(ltMDE_m1_array)[[3]][5] = "Between-group std. dev."

# define parameters to display in plots
parameters <- c(dimnames(ltMDE_m1_array)[[3]][1:5],
                groupName[1:42])

# Autocorrelation plots
autocor_plots(array = ltMDE_m1_array, 
              parsList = parameters)

# Trace plots for each parameter
trace_plots(array = ltMDE_m1_array,
            parsList = parameters)
```

### Model estimates

Age-standardized prevalence

```{r}
# Define population age distribution weights
age_weights <- c(0.138, 0.159, 0.246, 0.456)
# Note: these percentages come from Table 1
# 18-25 = 75,591 (13.8%)
# 26-34 = 48,242 (15.9%)
# 35-49 = 62,110 (24.6%)
# 50+   = 48,779 (45.6%)

# Generate age-standardized prevalence
ltMDE_prevalence <-
  age_adj_prevalence(model = ltMDE_m1, 
                     parameters =  c("b_Intercept", "b_age2634", 
                                     "b_age3549", "b_age50plus"),
                     age_weights = age_weights)
```

Model coefficients

```{r}
# Get ORs and CIs
ltMDE_m1_estimates <- ltMDE_m1_sum %>% 
  # keep rows with beta coefficients
  filter(row_number() %in% c(1:4)) %>% 
  mutate(
    parameter = c("Intecept",
                  "26-34",
                  "35-49",
                  "50+"),
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

## Full model: Fully-adjusted intersectional interaction

### Convergence diagnostics

```{r, eval = FALSE}
# histogram of r-hat values
mcmc_rhat_hist(rhat(ltMDE_m2)) +
  ggtitle("Histogram of model parameter r-hat values",
          "Full model: Intersectional interaction") +
  theme_clean() +
  theme(axis.title = element_text(face="bold")) +
  labs(y="Frequency", x="r-hat values")

# Rename parameter labels
ltMDE_m2_array <- as.array(ltMDE_m2)
dimnames(ltMDE_m2_array)[[3]][1] = "Intercept"
dimnames(ltMDE_m2_array)[[3]][2] = "age_2534"
dimnames(ltMDE_m2_array)[[3]][3] = "age_3549"
dimnames(ltMDE_m2_array)[[3]][4] = "age_50plus"
dimnames(ltMDE_m2_array)[[3]][5] = "Asian"
dimnames(ltMDE_m2_array)[[3]][6] = "Black"
dimnames(ltMDE_m2_array)[[3]][7] = "Hispanic/Latine"
dimnames(ltMDE_m2_array)[[3]][8] = "NAAN"
dimnames(ltMDE_m2_array)[[3]][9] = "NHPI"
dimnames(ltMDE_m2_array)[[3]][10] = "Multiracial"
dimnames(ltMDE_m2_array)[[3]][11] = "Woman"
dimnames(ltMDE_m2_array)[[3]][12] = "Gay/Lesbian"
dimnames(ltMDE_m2_array)[[3]][13] = "Bisexual"
dimnames(ltMDE_m2_array)[[3]][14] = "Between-group std. dev."

# define parameters to display in plots
parameters <- c(dimnames(ltMDE_m2_array)[[3]][1:14])

# Autocorrelation plots
autocor_plots(array = ltMDE_m2_array, 
              parsList = parameters)

# Trace plots
trace_plots(array = ltMDE_m2_array,
            parsList = parameters)
```

### Model estimates

```{r}
# Get ORs and CIs
ltMDE_m2_estimates = ltMDE_m2_sum %>% 
  # keep rows with beta coefficients
  filter(row_number() %in% c(1:13)) %>% 
  mutate(
    parameter = c("Intercept",
                  "26-34",
                  "35-49",
                  "50+",
                  "Asian",
                  "Black/African-American",
                  "Hispanic/Latine",
                  "Native American / Alaskan Native",
                  "Native Hawaiian / Pacific Islander",
                  "Multiracial",
                  "Woman",
                  "Gay/Lesbian",
                  "Bisexual"),
    OR = format(round(exp(mean),digits=2),nsmall=2),
    LB = format(round(exp(`2.5%`),digits=2),nsmall=2),
    UB = format(round(exp(`97.5%`),digits=2),nsmall=2),
    est = paste(OR," (",LB,", ",UB,")", sep="")
  ) %>% 
  select(parameter, est)
```

### Total, main, and intersectional effects

```{r}
ltMDE_effects <- 
  calculate_effects(model = ltMDE_m2,
                     parameters = c("b_Intercept", "b_age2634", 
                                    "b_age3549", "b_age50plus",
                                    "b_asian", "b_black",
                                    "b_hispanic", "b_naan",
                                    "b_nhpi", "b_multi",
                                    "b_woman", "b_gay",
                                    "b_bisexual"),
                     age_weights = age_weights)


# Combine all estimated effects into a single table
ltMDE_effect_summary <-
  # start with total effects
  ltMDE_effects$total_sum %>% 
  select(`LT Total Effects (95% CI)` = Combined) %>% 
  # then add main effects
  cbind(., 
        ltMDE_effects$main_sum %>% 
          select(`LT Main Effects (95% CI)` = Combined)) %>% 
  # then add intersectional effects
  cbind(., 
        ltMDE_effects$intersectional_sum %>% 
          select(`LT % Point Diff.` = Combined)) %>% 
  cbind(.,
        ltMDE_effects$intersectional_pct_sum %>% 
          select(`LT % Diff.` = Combined))  %>%  
  # add intersectional group names and group IDs
  cbind(`Intersectional Group` = groupName, `Group ID` = c(1:42), .)
```

# Save output

This output will be used to create table and figures in the file `05_tables_figures.Rmd`.

```{r}
# prevalence estimates
saveRDS(pyMDE_prevalence,
        "03_tables_figures/model_estimates/sensitivity/pyMDE_prevalence.RDS")
saveRDS(ltMDE_prevalence,
        "03_tables_figures/model_estimates/sensitivity/ltMDE_prevalence.RDS")

# model fit summaries
saveRDS(pyMDE_m1_sum,
        "03_tables_figures/model_estimates/sensitivity/pyMDE_m1_sum.RDS")
saveRDS(pyMDE_m2_sum,
        "03_tables_figures/model_estimates/sensitivity/pyMDE_m2_sum.RDS")
saveRDS(ltMDE_m1_sum,
        "03_tables_figures/model_estimates/sensitivity/ltMDE_m1_sum.RDS")
saveRDS(ltMDE_m2_sum,
        "03_tables_figures/model_estimates/sensitivity/ltMDE_m2_sum.RDS")
saveRDS(pyMDE_m1_estimates,
        "03_tables_figures/model_estimates/sensitivity/pyMDE_m1_estimates.RDS")
saveRDS(pyMDE_m2_estimates,
        "03_tables_figures/model_estimates/sensitivity/pyMDE_m2_estimates.RDS")
saveRDS(ltMDE_m1_estimates,
        "03_tables_figures/model_estimates/sensitivity/ltMDE_m1_estimates.RDS")
saveRDS(ltMDE_m2_estimates,
        "03_tables_figures/model_estimates/sensitivity/ltMDE_m2_estimates.RDS")
# variance
saveRDS(pyMDE_var,
        "03_tables_figures/model_estimates/sensitivity/pyMDE_var.RDS")
saveRDS(ltMDE_var,
        "03_tables_figures/model_estimates/sensitivity/ltMDE_var.RDS")

# intersectional effects (excess/reduced prevalence)
saveRDS(pyMDE_effect_summary,
        "03_tables_figures/model_estimates/sensitivity/pyMDE_effect_summary.RDS")
saveRDS(ltMDE_effect_summary,
        "03_tables_figures/model_estimates/sensitivity/ltMDE_effect_summary.RDS")
```
